cleanup: added in helper class for getters and setters as makes it tidier
